requirements.txt
azure-cognitiveservices-speech==1.38.0
boto3==1.40.0
python-dotenv==1.0.0
typer==0.12.0
requests==2.31.0
google-cloud-speech==2.27.0
